+++ lib/rdf/serializations/trig/encoder.ex
@@ -11,10 +11,23 @@ defmodule RDF.TriG.Encoder do
 
   ## Options
 
-  - `:only`: Allows to specify which parts of a TriG document should be generated.
-    Possible values: `:base`, `:prefixes`, `:directives` (means the same as `[:base, :prefixes]`),
-    `:default_graph`, `:named_graphs`, `:graphs` (means the same as `[:default_graph, :named_graphs]`)
-    or a list with any combination of these values.
+  - `:content`: Allows specifying the content and structure of the Turtle document
+    to be rendered and defining which parts should be generated in which order.
+    This option accepts lists of the values `:base`, `:prefixes`, `:default_graph` and `:named_graphs`.
+    You can also use `:directives` to specify `[:base, :prefixes]` and `:graphs`
+    to specify `[:default_graph, :named_graphs]` as a group.
+    Additionally, arbitrary strings can be included at desired positions to customize
+    the document.
+
+        RDF.TriG.write_string(dataset, content: [
+          "# === HEADER ===\\n\\n",
+          :directives,
+          "\\n# === NAMED GRAPHS ===\\n\\n",
+          :named_graphs
+          "\\n# === DEFAULT GRAPH ===\\n\\n",
+          :default_graph
+        ])
+
   - `:prefixes`: Allows to specify the prefixes to be used as a `RDF.PrefixMap` or
     anything from which a `RDF.PrefixMap` can be created with `RDF.PrefixMap.new/1`.
     If not specified the prefixes from all the graphs of the given dataset are used