+++ CHANGELOG.md
@@ -11,6 +11,8 @@ This project adheres to [Semantic Versioning](http://semver.org/) and
 ### Added
 
 - Decoder and encoder of the TriG serialization language. 
+- Capability to add custom content in the `RDF.Turtle.Encoder` (and new `RDF.TriG.Encoder`)
+  with the `:content` option.
 - `RDF.Dataset.update_all_graphs/2` to apply a function on all graphs of a dataset.
 - `RDF.Dataset.named_graphs/1` to get a list of all named graphs of a dataset.
 - `RDF.Dataset.graph_names/1` to get a list of all graph names of a dataset.
@@ -21,6 +23,8 @@ This project adheres to [Semantic Versioning](http://semver.org/) and
 ### Changed
 
 - Update to change in N-Triples and N-Quads specs disallowing colons in bnode labels
+- Rename `:only` option of `RDF.Turtle.Encoder` to `:content` to reflect the enhanced 
+  capabilities.
 
 ### Fixed